requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.4
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.2.2
lightning-utilities==0.11.2
lit==18.1.3
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.4.0
prettytable==3.10.2
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==1.45.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.1.1
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.39.3
triton==2.1.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
wcwidth==0.2.13
yarl==1.9.4
zipp==3.20.2
